requirements.txt
argparse
soundfile
ebooklib
beautifulsoup4
kokoro-onnx
pydub
torch
pick
onnxruntime